+++ index.ts
@@ -1,3 +1,8 @@
 export * from './src/currency-mask.directive';
 export * from './src/currency-mask.module';
-export { CurrencyMaskConfig, CurrencyMaskInputMode } from './src/currency-mask.config';
+
+export {
+  CurrencyMaskConfig,
+  CurrencyMaskInputMode,
+  CURRENCY_MASK_CONFIG,
+} from './src/currency-mask.config';