+++ src/input.service.ts
@@ -121,8 +121,13 @@ export class InputService {
         // Ensure max is at least as large as min.
         max = (this.isNullOrUndefined(max) || this.isNullOrUndefined(min)) ? max : Math.max(max, min);
 
+        // Ensure precision number works well with more than 2 digits
+        // 23 / 100... 233 / 1000 and so on
+        const divideBy = Number('1'.padEnd(precision + 1, '0'));
+        
         // Restrict to the min and max values.
-        let newValue = integerValue + (decimalValue / 100);
+        let newValue = integerValue + (decimalValue / divideBy);
+
         newValue = isNegative ? -newValue : newValue;
         if (!this.isNullOrUndefined(max) && newValue > max) {
             return this.applyMask(true, max + '');
@@ -227,6 +232,10 @@ export class InputService {
 
         let shiftSelection = 0;
         let insertChars = '';
+        
+        const isCursorInDecimals = decimalIndex < selectionEnd;
+        const isCursorImmediatelyAfterDecimalPoint = decimalIndex + 1 === selectionEnd;
+        
         if (selectionEnd === selectionStart) {
             if (keyCode == 8) {
                 if (selectionStart <= prefix.length) {
@@ -240,8 +249,12 @@ export class InputService {
                 }
 
                 // In natural mode, jump backwards when in decimal portion of number.
-                if (inputMode === CurrencyMaskInputMode.NATURAL && decimalIndex < selectionEnd) {
+                if (inputMode === CurrencyMaskInputMode.NATURAL && isCursorInDecimals) {
                     shiftSelection = -1;
+                    // when removing a single whole number, replace it with 0
+                    if (isCursorImmediatelyAfterDecimalPoint && this.value < 10 && this.value > -10) {
+                        insertChars += '0';
+                    }
                 }
             } else if (keyCode == 46 || keyCode == 63272) {
                 if (selectionStart === suffixStart) {